MySolution
I'm not an expert on workign with Git or using shell scripts. I've spent a lot of time trying to figure out what I can. Hoping this satisfies some requirments. I'm very excited to have the opportunity to work with Insight Fellows. Nevertheless, I appreciate this project to allow me to learn that I'm capable.